refactor: streamline exports in RichText editor components and update mention types
