+++ packages/react/src/experimental/RichText/CoreEditor/Extensions/Mention/types.tsx
@@ -1,7 +1,3 @@
-import { MentionedUser } from "../../../RichTextEditor/utils/types"
-
-export type { MentionedUser }
-
 export interface MentionNodeAttrs {
   id: string
   label: string
@@ -17,4 +13,18 @@ export interface MentionItemComponentProps {
   item: MentionedUser
   index: number
   selected: boolean
+}
+
+export type MentionedUser = {
+  id: number
+  label: string
+  image_url?: string
+  href?: string
+}
+
+export type MentionsConfig = {
+  onMentionQueryStringChanged?: (
+    queryString: string
+  ) => Promise<MentionedUser[]> | undefined
+  users: MentionedUser[]
 }
