+++ packages/react/src/experimental/RichText/CoreEditor/index.tsx
@@ -1,0 +1,3 @@
+export * from "./BubbleMenu"
+export * from "./Extensions"
+export * from "./Toolbar"
