+++ packages/react/src/experimental/RichText/RichTextEditor/Footer/ActionsMenu/index.tsx
@@ -4,7 +4,7 @@ import {
   OneDropdownButtonItem,
 } from "@/components/Actions/OneDropdownButton"
 import { Switch } from "@/experimental/Forms/Fields/Switch"
-import { ToolbarDivider } from "@/experimental/RichText/CoreEditor/Toolbar/ToolbarDivider"
+import { ToolbarDivider } from "@/experimental/RichText/CoreEditor"
 import { primaryActionType, secondaryActionType } from "../../utils/types"
 
 interface ActionsMenuProps {
